[CI] Add Release Notes Automation
Add new workflow to generate release notes. This workflow can be run
manually via workflow dispatch, but will also run automatically upon
creating a release. The workflow dispatch can be used for testing,
previewing release notes for a potential release, or for generating
release notes for past releases for when the automation has a bug or
didn't exist.